+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './object.utils';
+export * from './retry';
 export * from './table-transforms';